Fix for mispredict when using tools on recycler. (#43766)
* predict

* fix
